refactor to separate files and pull out account api
