+++ src/utils/api_manager.js
@@ -1,308 +1,18 @@
 import Vue from 'vue';
-import { i18n } from '@/utils'
-import router from '@/router/';
 import store from '@/store/';
 import emojione from 'emojione';
 import * as firebase from 'firebase';
 import ImageCompressor from '@xkeshi/image-compressor';
-import ReconnectingWebsocket from 'reconnecting-websocket';
 
 import { Util, Url, Crypto, SessionCache, Platform } from '@/utils/'
+import { Account, Conversations, Stream } from '@/utils/api/'
 
 export default class Api {
 
-    static stream = class {
-        constructor() {
-            this.socket = null;
-            this.open();
+    static stream = Stream
 
-            this.has_disconnected = false;
-            this.disconnected_timeout = null;
-        }
-
-        /**
-         * Open reconnecting websocket.
-         */
-        open() {
-            const this_ = this;
-
-            this.socket = new ReconnectingWebsocket(Url.get('websocket') + Url.getAccountParam());
-
-            this.socket.addEventListener('open', () => {
-                if (this.has_disconnected) {
-                    store.state.msgbus.$emit('refresh-btn');
-                    Util.snackbar(i18n.t('api.back'));
-                }
-
-                if (this.disconnected_timeout != null) {
-                    clearTimeout(this.disconnected_timeout);
-                    this.disconnected_timeout = null;
-                }
-
-                this.has_disconnected = false;
-
-                const subscribe = JSON.stringify({
-                    "command": "subscribe",
-                    "identifier": JSON.stringify({
-                        "channel": "NotificationsChannel"
-                    })
-                });
-
-                this.socket.send(subscribe);
-            });
-
-            this.socket.addEventListener('message', (e) => this.handleMessage(e));
-
-            this.socket.addEventListener('close', (e) => {
-                if (e.wasClean || e.code == 1001) // If not an error, ignore
-                    return
-
-                // if the websocket reconnects within 5 seconds, we won't show anything to the user
-                // if it takes longer than that, then we will notify the user that we are
-                // trying to reconnect.
-                this.disconnected_timeout = setTimeout(() => {
-                    if (!this.has_disconnected)
-                        Util.snackbar(i18n.t('api.disconnected'));
-
-                    this.has_disconnected = true;
-                }, 5 * 1000);
-            });
-        }
-
-        /**
-         * Perminently close socket
-         */
-        close() {
-            this.socket.close(1000, '', { keepClosed: true });
-        }
-
-        /**
-         * Handle incoming socket data
-         * @param e - socket event
-         */
-        handleMessage(e) {
-            if (e.data.indexOf("ping") != -1) { // Is keep alive event
-                // Store last ping to maintain data connection
-                store.commit('last_ping', Date.now() / 1000 >> 0);
-                return;
-            }
-
-            // Parse out JSON
-            const json = JSON.parse(e.data);
-
-            // Ignore bad messages
-            if (typeof json.message == "undefined")
-                return;
-
-            const operation = json.message.operation;
-
-            // Parse any emojis
-            if (typeof json.message.content.data != "undefined")
-                json.message.content.data = emojione.unicodeToImage(
-                    json.message.content.data
-                );
-
-
-            if (operation == "added_message") {
-                let message = json.message.content;
-                message.message_from = message.from;
-                message = Crypto.decryptMessage(message);
-
-                this.notify(message);
-
-                SessionCache.cacheMessage(message);
-                SessionCache.updateConversation(message);
-
-                store.state.msgbus.$emit('newMessage', message);
-            } else if (operation == "removed_message") {
-                let message = json.message.content;
-
-                store.state.msgbus.$emit('deletedMessage', message.id);
-            } else if (operation == "read_conversation") {
-                const id = json.message.content.id;
-
-                SessionCache.readConversation('index_public_unarchived');
-                SessionCache.readConversation('index_archived');
-
-                store.state.msgbus.$emit('conversationRead', id);
-            } else if (operation == "update_message_type") {
-                const id = json.message.content.id;
-                const message_type = json.message.content.message_type;
-
-                SessionCache.updateMessageType(id, message_type);
-                store.state.msgbus.$emit('updateMessageType-' + id, { message_type });
-            } else if (operation == "added_conversation") {
-                const id = json.message.content.id;
-
-                SessionCache.invalidateConversations('index_public_unarchived');
-                store.state.msgbus.$emit('addedConversation', { id });
-            } else if (operation == "removed_conversation") {
-                const id = json.message.content.id;
-
-                SessionCache.removeConversation(id, 'index_public_unarchived');
-                SessionCache.removeConversation(id, 'index_archived');
-                SessionCache.removeConversation(id, 'index_private');
-                store.state.msgbus.$emit('removedConversation', { id });
-            } else if (operation == "archive_conversation") {
-                const id = json.message.content.id;
-
-                if (json.message.content.archive) {
-                    SessionCache.removeConversation(id, 'index_public_unarchived');
-                    SessionCache.invalidateConversations('index_archived');
-                } else {
-                    SessionCache.removeConversation(id, 'index_archived');
-                    SessionCache.invalidateConversations('index_public_unarchived');
-                }
-
-                store.state.msgbus.$emit('removedConversation', { id });
-            }
-        }
-
-        /**
-         * Submit notification for message
-         * @param message  - message object
-         */
-        notify(message) {
-            if (Notification.permission != "granted" && !store.state.notifications)
-                return
-
-            if (message.type != 0)
-                return;
-
-            if (!Platform.isWebsite()) {
-                return;
-            }
-
-            // fetch through the API instead of the session cache, since the session
-            // cache doesn't know about the mute/private settings
-            Api.fetchConversation(message.conversation_id).then(conversation => {
-                if (conversation == null || conversation.mute) {
-                    return;
-                }
-
-                const title = conversation.title;
-                const snippet = conversation.private_notifications ? "" : Util.generateSnippet(message);
-
-                const link = "/thread/" + message.conversation_id;
-
-                const notification = new Notification(title, {
-                    icon: '/static/images/android-desktop.png',
-                    body: snippet
-                });
-
-                notification.onclick = () => {
-                    window.focus()
-                    router.push(link);
-                }
-            })
-        }
-    }
-
-    static login(username, password) {
-        const promise = new Promise((resolve, reject) => {
-            const constructed_url = Url.get('login')
-            const request = {
-                username,
-                password
-            };
-
-            Vue.http.post(constructed_url, request, { 'Content-Type': 'application/json' })
-                .then((response) => resolve(response))
-                .catch((error) => reject(error));
-
-        });
-
-        return promise
-    }
-
-    static conversations = {
-        getList: (index, folderId) => {
-            if (typeof index == 'undefined') {
-                index = "index_public_unarchived"
-            }
-
-            if (index == 'folder') {
-                index = index + "/" + folderId;
-            }
-
-            let constructed_url =
-                Url.get('conversations') + index + Url.getAccountParam() + "&limit=75"
-
-            const promise = new Promise((resolve, reject) => {
-                if (!SessionCache.hasConversations(index)) {
-                    Vue.http.get(constructed_url)
-                        .then(response => {
-                            response = response.data
-
-                            if (response != null) {
-                                // Decrypt Conversations items
-                                for (let i = 0; i < response.length; i++) {
-                                    const convo = Crypto.decryptConversation(response[i]);
-                                    if (convo != null)
-                                        response[i] = convo;
-                                }
-
-                                SessionCache.putConversations(response, index);
-                            }
-
-                            resolve(response); // Resolve response
-                        })
-                        .catch(response => Api.rejectHandler(response, reject));
-                } else {
-                    resolve(SessionCache.getConversations(index));
-                }
-            });
-
-            return promise
-        },
-        getById: (id) => {
-            let constructed_url = Url.get('conversation') + id + Url.getAccountParam()
-            const promise = new Promise((resolve, reject) => {
-                Vue.http.get(constructed_url)
-                    .then(response => {
-                        response = Crypto.decryptConversation(response.data);
-                        if (response != null) {
-                            resolve(response);
-                        }
-                    })
-                    .catch(response => Api.rejectHandler(response, reject));
-            });
-
-            return promise
-        },
-        update: (conversation_id, params) => {
-            let constructed_url = Url.get('update_conversation') + conversation_id;
-            Vue.http.post(constructed_url, params, { 'Content-Type': 'application/json' })
-                .catch(response => console.log(response));
-        },
-        read: (conversation_id) => {
-            // Read conversation
-            let constructed_url = Url.get('read') + conversation_id + Url.getAccountParam();
-            Vue.http.post(constructed_url)
-                .catch((e) => Api.rejectHandler(e));
-
-            // Dismiss notifiction
-            constructed_url = Url.get('dismiss') + Url.getAccountParam()
-                + "&id=" + conversation_id;
-            Vue.http.post(constructed_url);
-        },
-        archive: (conversation_id, archive) => {
-            let constructed_url;
-
-            if (archive) {
-                constructed_url = Url.get('archive')
-            } else {
-                constructed_url = Url.get('unarchive')
-            }
-
-            constructed_url += conversation_id + Url.getAccountParam();
-            Vue.http.post(constructed_url);
-        },
-        delete: (conversation_id) => {
-            let constructed_url = Url.get('delete') + conversation_id + Url.getAccountParam();
-            Vue.http.post(constructed_url);
-        }
-    }
+    static account = Account
+    static conversations = Conversations
 
     static fetchThread(conversation_id, offset = 0) {
         const limit = 70;
@@ -639,16 +349,7 @@ export default class Api {
         return promise;
     }
 
-    static fetchAccount() {
-        const constructed_url = Url.get('account_stats') + Url.getAccountParam();
-        const promise = new Promise((resolve, reject) => {
-            Vue.http.get(constructed_url)
-                .then((response) => resolve(response))
-                .catch((error) => reject(error));
-        });
-
-        return promise
-    }
+    
 
     static fetchDrafts() {
         let constructed_url = Url.get('drafts') + Url.getAccountParam();
@@ -747,57 +448,6 @@ export default class Api {
     static removeAutoReply(id) {
         let constructed_url = Url.get('remove_auto_reply') + id + Url.getAccountParam();
         Vue.http.post(constructed_url);
-    }
-
-    static fetchSettings() {
-        let constructed_url = Url.get('settings') + Url.getAccountParam();
-        const promise = new Promise((resolve, reject) => {
-            Vue.http.get(constructed_url)
-                .then(response => {
-                    Api.processSettings(response);
-                    resolve(response);
-
-                })
-                .catch(response => Api.rejectHandler(response, reject));
-        });
-
-        return promise
-    }
-
-    static processSettings(response) {
-        response = response.data
-
-        const colors = {
-            'default': Util.expandColor(response.color),
-            'dark': Util.expandColor(response.color_dark),
-            'accent': Util.expandColor(response.color_accent),
-        };
-
-        if (colors.default == 'rgba(255,255,255,1)' && colors.dark == 'rgba(255,255,255,1)' && colors.accent == 'rgba(255,255,255,1)') {
-            colors.default = store.theme_global_default;
-            colors.dark = store.theme_global_dark;
-            colors.accent = store.theme_global_default_accent;
-        }
-
-        store.commit('theme_base', response.base_theme);
-        store.commit('theme_use_global', response.use_global_theme);
-        store.commit('theme_apply_appbar_color', response.apply_primary_color_to_toolbar);
-        store.commit('theme_global', colors);
-        store.commit('colors', colors);
-    }
-
-    static updateSetting(setting, type, value) {
-        let constructed_url = Url.get("update_setting") +
-            "?pref=" + setting
-            + "&type=" + type
-            + "&value=" + value;
-
-        const promise = new Promise((resolve, reject) => {
-            Vue.http.post(constructed_url, Url.getAccountPayload(),
-                { 'Content-Type': 'application/json' })
-                .then(response => resolve(true))
-                .catch(response => Api.rejectHandler(resposne, reject));
-        });
     }
 
     static fetchContacts() {
